backend/requirements.txt
django 
djangorestframework 
django-cors-headers 
langchain 
langchain-community 
langchain-postgres
langchain-google-genai
psycopg2-binary